Fixed makefile

+++ Makefile
@@ -3,7 +3,7 @@ OBJ = $(subst .c,.o,$(subst src,work,$(SRC)))
 LIBS = -pthread -pie -lrt $(shell pkg-config --libs gnutls)
 CFLAGS := -O2 -pipe -Wall -Wpedantic -Werror
 CFLAGS += -Isrc/include -fpie $(shell pkg-config --cflags gnutls)
-INSTALLDIR := /usr/bin
+INSTALLDIR := /usr/sbin
 OUT = swebs
 
 build/$(OUT): $(OBJ)
@@ -14,8 +14,8 @@ work/%.o: src/%.c $(wildcard src/include/*.h)
 
 install: build/$(OUT)
 	cp build/$(OUT) $(INSTALLDIR)/$(OUT)
-	useradd -M swebs
+	if ! id swebs >> /dev/null 2>&1; then useradd -M swebs; fi
 
 uninstall: $(INSTALLDIR)/$(OUT)
 	rm $(INSTALLDIR)/$(OUT)
-	userdel swebs
+	if id swebs >> /dev/null 2>&1; then userdel swebs; fi
